pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "discord-llm-translator"
version = "0.1.0"
description = "Discord bot for automatic translation using LLMs via OpenRouter.ai"
readme = "README.md"
license = "AGPL-3.0"
requires-python = ">=3.11"
authors = [
    { name = "DiscordLLMTranslatorBot contributors" }
]
keywords = ["discord", "bot", "translation", "llm", "openrouter"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

dependencies = [
    "discord.py>=2.4.0",
    "aiohttp>=3.9.0",
    "langdetect>=1.0.9",
    "tomli>=2.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pyright>=1.1.370",
]

[project.scripts]
discord-translator = "discord_llm_translator.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/discord_llm_translator"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
    "RUF",
]
ignore = [
    "E501",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "aiohttp.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "discord.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "langdetect.*"
ignore_missing_imports = true
